Add annotations for API call arguments
This allows users to transform incoming variables – and more importantly, change them, presenting a nicer call interface.
For `nationalrail` this will be useful to turn `crs` into `crs_or_station_name`, with a function that automatically does station name lookup.

An inliner checks if an 'annotator' is a simple return; if so it inlines the function.

@+ feels like a hacky syntax, but this entire project is hacky; it's cool.

The functionality is mostly there to transform properties, too, via creation of a MyClass._from_api.
However, as these properties might be deeply embedded in a dict, it needs a bit of work to recursively figure out where to do this.