Handle call annotators far, far better
ast_syntax_error is adorable. It got me curious a little as to how Python does it internally; turns out ast.AST.end_offset was in 3.8, long before it was used in 3.10 for nice messages.

Hope this is helpful, anyway, and not a total headache. You can never tell with AST stuff, can you?